Removed all individual components from request process
